update tests for new api
