+++ test/dispatch-test.js
@@ -6,41 +6,39 @@ var me = require('../');
 var provide = me.reduxProvide;
 var dispatch = me.reduxDispatch;
 
-tape('selection.dispatchOn(type, listener) dispatches on event', function(test) {
+tape('selection.on(type, dispatch(actionCreator)) dispatches on event', function(test) {
   var document = jsdom('<div></div>');
   var div = document.querySelector('div');
   var actions = [];
-  var store = {
-    dispatch: function (action) { actions.push(action);}
-  };
   var sel = d3.select(document.body)
-    .call(provide(store))
+    .call(provide({ dispatch: function (action) { actions.push(action);} }))
     .select('div')
     .on('click', dispatch(function () { return 42; }));
+
   sel.dispatch('click');
 
   test.deepEqual(actions, [42]);
   test.end();
 });
 
-tape('selection.dispatchOn(type, listener) passes the listener data, index and group', function(test) {
-  var document = jsdom('<parent id="one"><child id="three"></child><child id="four"></child></parent><parent id="two"><child id="five"></child></parent>'),
-    one = document.querySelector('#one'),
-    two = document.querySelector('#two'),
-    three = document.querySelector('#three'),
-    four = document.querySelector('#four'),
-    five = document.querySelector('#five'),
-    results = [];
-
+tape('selection.on(type, dispatch(actionCreator) passes the actionCreator data, index and group', function(test) {
+  var document = jsdom('<parent id="one"><child id="three"></child><child id="four"></child></parent><parent id="two"><child id="five"></child></parent>');
+  var one = document.querySelector('#one');
+  var two = document.querySelector('#two');
+  var three = document.querySelector('#three');
+  var four = document.querySelector('#four');
+  var five = document.querySelector('#five');
+  var results = [];
   var selection = d3.selectAll([one, two])
     .call(provide({ dispatch: function () {} }))
     .datum(function(d, i) { return 'parent-' + i; })
     .selectAll('child')
     .data(function(d, i) { return [0, 1].map(function(j) { return 'child-' + i + '-' + j; }); })
     .on('foo', dispatch(function(d, i, nodes) { results.push([this, d, i, nodes]); }));
-
   test.deepEqual(results, []);
+
   selection.dispatch('foo');
+
   test.deepEqual(results, [
     [three, 'child-0-0', 0, [three, four]],
     [four, 'child-0-1', 1, [three, four]],
